Route console.log messages properly in pmloglib
:Release Notes:
Make sure console messages make it to the node-inspector

:Detailed Notes:
When using the node-pmloglib module to route console messages to the log
file, we were accidentally disconnecting them from the Node console,
which caused the messages to not show up in node-inspector. Fixed that
so that messages go to *both* locations.

:Testing Performed:
Tested that console.log messages end up in both places.

:QA Notes:

:Issues Addressed:
[ANG-244] Node services console.log needs to be properly directed

Open-webOS-DCO-1.0-Signed-off-by: Mark Bessey <mark.bessey@lge.com>

Change-Id: I60d70ea282122fe8ded10a8f68495ff177d040b3
Reviewed-on: https://g2g.palm.com/4706
Reviewed-by: DCO Verification
Reviewed-by: Ben Combee <ben.combee@lge.com>
Reviewed-by: Mark Bessey <mark.bessey@lge.com>
Tested-by: Mark Bessey <mark.bessey@lge.com>